refactor: Remove unused HorizontalScrollLabel component
The HorizontalScrollLabel component and its associated CSS and constants
were removed as they are no longer being used in any sections of the
application.

Additionally, a minor adjustment was made to the `--contacts-footer-width`
variable in `ContactsFooter.css` to reduce its width slightly.

The media query in `UsedCarDrawer.css` was also made more specific by
adding `and (orientation: portrait)`.